changed config generator script

+++ yace-config-generator.sh
@@ -15,9 +15,10 @@ then
       echo "File exists and is not empty"
       cat /root/dockprom/yace/config.yml >> $TEMP_FILE
       for i in $LIST; do
-        if [[ " ${array[*]} " =~ $i ]]; then
+        if [[ "${array[*]} " =~ $i ]]; then
           cat /root/dockprom/yace/$i.yml >> $TEMP_FILE;
-        fi
+          echo "file $i added"
+        fi  
       done
     else
       echo "File exists but empty"
@@ -25,13 +26,15 @@ then
 else
     echo "File not exists"
 fi
-if cmp -s "$TEMP_FILE" "$YACE_CONF_FILE"; then
- echo "Files are different therefore updating configuration"   
- mv $YACE_CONF_FILE /etc/yace/config_$(date +"%Y-%m-%d-%H-%M-%S").yml
- cp $TEMP_FILE $YACE_CONF_FILE
- rm -rf $TEMP_FILE
- echo "Updated yace config for new aws resources"
- systemctl restart yace.service
-else
- echo "No new resources have been added"
+if [ -f ${YACE_CONF_FILE} ]; then
+  if cmp -s "$TEMP_FILE" "$YACE_CONF_FILE"; then
+    echo "Files are different therefore updating configuration"   
+    mv $YACE_CONF_FILE /etc/yace/config_$(date +"%Y-%m-%d-%H-%M-%S").yml
+    cp $TEMP_FILE $YACE_CONF_FILE
+    rm -rf $TEMP_FILE
+    echo "Updated yace config for new aws resources"
+    systemctl restart yace.service
+  else
+    echo "No new resources have been added"
+  fi
 fi
